code/backend/requirements.txt
# Core
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0

# Database
pymongo==4.3.3
redis==4.5.5
pika==1.3.2

# Auth
python-keycloak==2.15.0
python-jose==3.3.0
passlib==1.7.4

# Storage
minio==7.1.15

# Monitoring
elastic-apm==6.23.0

# Optional (uncomment if needed)
# gunicorn==20.1.0  # For production
# httpx==0.24.1     # For async HTTP requests  